修改均值阈值滤波的均值计算，从计算8组的均值改为计算10组的均值

+++ 单片机代码/0911_QiDong/Inc/usart.h
@@ -45,7 +45,7 @@ extern int LCD_Upper_Jiaozhun[6];
 extern int LCD_Lower_Jiaozhun[6];
 extern int LCD_Upper_GongCha[6];
 extern int LCD_Lower_GongCha[6];
-extern float K[6];
+extern int K[6];
 extern int LCD_high[6];
 extern int LCD_low[6];
